// dotnet/openapi/model/clientCompanyModelServiceResponse.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ClientCompanyModel } from './clientCompanyModel';


export interface ClientCompanyModelServiceResponse { 
    message?: string | null;
    statusCode?: number;
    data?: ClientCompanyModel;
}