// dotnet/openapi/model/updateClientCompanyCommand.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateClientCompanyCommand { 
    companyName?: string | null;
    contactPersonName?: string | null;
    contactEmail?: string | null;
    contactPhone?: string | null;
    designation?: string | null;
    companyWebsite?: string | null;
    linkedInProfile?: string | null;
    location?: string | null;
    notes?: string | null;
    id?: string;
}